add data utils folder move mock data
