+++ assets/src/dashboard/dataUtils/formattedStoriesArray.js
