+++ assets/src/dashboard/dataUtils/formattedTemplatesArray.js
